fix leave prompt error

+++ src/ToktokWallet/screens/KYCScreens/ToktokWalletVerification/Components/Confirm/index.js
@@ -63,7 +63,7 @@ export const Confirm = connect(mapStateToProps, mapDispatchToProps)(({session})=
             // removeCacheImages({
             //     VerifyUserData
             // })
-            RNFS.unlink(RNFS.CachesDirectoryPath)
+            if(RNFS.CachesDirectoryPath) RNFS.unlink(RNFS.CachesDirectoryPath)
             if(result.status == 2){
                 navigation.pop(2)
                 navigation.navigate("ToktokWalletVerifyResult")
@@ -142,7 +142,7 @@ export const Confirm = connect(mapStateToProps, mapDispatchToProps)(({session})=
         })
         : null
 
-        // RNFS.unlink(RNFS.CachesDirectoryPath).then(()=>{
+        // if(RNFS.CachesDirectoryPath) RNFS.unlink(RNFS.CachesDirectoryPath).then(()=>{
         //     console.log("Deleted")
         // }).catch(err=>console.log(err))
         // RNFS.unlink(RNFS.TemporaryDirectoryPath)
